requirements.txt
google-api-python-client==2.149.0
google-auth==2.35.0
pandas==2.2.3
geotext==0.4.0
openai==1.51.2
python-dotenv==1.0.1
selenium==4.25.0
beautifulsoup4==4.12.3
webdriver-manager==4.0.2
requests==2.32.3
gspread==6.1.4
oauth2client==4.1.3